GithubFetcher/requirements.txt

fastapi
httpx
loguru
pytest
tenacity
python-dotenv
uvicorn
